# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Mobile Shop POS API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  category: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  date: string;
  created_at?: string;
}