# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Mobile Shop POS API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  username: string;
  name: string;
  role: UserRole;
  created_at?: string;
}